Cnam-LMSSCwav2vec2-french-phonemizer/requirements.txt
streamlit
numpy
phonemizer
torch
torchaudio
transformers
python-Levenshtein
Pillow
